model_servers/object_detection_server/requirements_server.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
Pillow==11.1.0

